fix(dlsp): Fix misleading on plan loading
To make sure that a single plan file is loaded, `plan.New` has been rework to
invert loading process, it will first try as a file and then as a directory.

This way, if the loaded file include an external definition from another file,
it will load it as a directory.
The issue was that it was possible to load multiple plan as a directory because
they were merged into a single one.

Resolves: #10
Signed-off-by: Vasek - Tom C <tom.chauveau@epitech.eu>