Support import aliases
It doesn't seem possible to get package aliases from cue's build
instances. They only seem to be used internally in the adt and not
exported in the public API.

Therefore the patch relies on the ast to record package aliases by
loaded file. When the selector of a definition isn't found we make some
extra calls to see if it's a package alias in the current file and find
the associated instance.

Closes #79